Refined Projectile Classes

+++ potatoes/src/main/java/com/deco2800/potatoes/entities/enemies/MeleeAttack.java
@@ -93,7 +93,7 @@ public class MeleeAttack extends Projectile {
     }
 
     @Override
-    public int rotateAngle() {
+    public float rotateAngle() {
         return rotateAngle;
     }
